refactor: share downsampled bitmap loading for wallpaper
